Created timer, console and screen control syscalls.
Though those will be replaced when we actually build a X-window system,
they come in handy when implementing console applications. For example,
console syscall can control the buffering of the Kernel console and
screen syscall can help the user programs to have an access to the
kernel graphic memory.

For the testing, I created snake game (which is incomplete thought)
which demonstrates the Non blocking & No buffering IO, graphic memory
access and timer syscall altoghter.